Big Refactor
Refactor isn’t finished yet, but a couple of notes:

- removing coffee script entirely. Coffee script as gone the way of the
dinosaur, so it’s all ES6 from this point forward
- refactoring from the ground-up with caching in mind. The simulation
will integrate as many frames (ticks?) as it can in a single update.
For small simulations this means we’ll be able to jump forward and
backward in time. We’ll be able to cache something complex, then go
back and watch it in real time.
- Going to refactor the react ui, likely drop bootstrap and do a better
job of the mouse interactions.