Drastically increased maximum cache size
Now, the max number of ticks changes based on the amount of ticks each
body has cached.

Therefor, if a body is destroyed and stops caching ticks, the max cache
size increases.

A simulation of thousands of bodies might only be able to last for a
couple of minutes if those bodies all persist, but as they are
destroyed by collisions, it ends up being quite long.

Also added the ability for extreme speed playback in the UI